Refactor: Extracting function out from the huge updateFormLayoutSagas.ts (it contained functions for not just the form layout, but repeating groups and file uploaders)
